requirements.txt
streamlit
ollama
pytesseract
Pillow
PyMuPDF
pdf2image
python-docx